Use fallback project name rather than "pulum"
Fixes https://github.com/pulumi/pulumi/issues/13946.

This also moves some of the logic from sdk/go/common to pkg/ as it's
only used be `new` and `up.
