+++ changelog/pending/20230914--cli-new--pulumi-new-no-longer-defaults-to-a-project-name-of-pulum-if-ran-in-a-folder-called-pulumi.yaml
@@ -1,0 +1,4 @@
+changes:
+- type: fix
+  scope: cli/new
+  description: `pulumi new` no longer defaults to a project name of "pulum" if ran in a folder called "pulumi".
